ear/meteo/requirements.txt
requests==2.31.0
gTTS==2.3.2
pygame==2.5.1
python-dotenv==1.0.0
python-dateutil==2.8.2